// apps/web/tokens/generated/class-maps.ts
export const pbMap: BaseClassMap<Spacing> = {
  "0": "pb-0",
  "100": "pb-100",
  "200": "pb-200",
  "300": "pb-300",
  "400": "pb-400",
  "500": "pb-500",
  "600": "pb-600",
  "700": "pb-700",
  "800": "pb-800",
  "900": "pb-900",
};

export const plMap: BaseClassMap<Spacing> = {
  "0": "pl-0",
  "100": "pl-100",
  "200": "pl-200",
  "300": "pl-300",
  "400": "pl-400",
  "500": "pl-500",
  "600": "pl-600",
  "700": "pl-700",
  "800": "pl-800",
  "900": "pl-900",
};

export const mMap: ResponsiveClassMap<Spacing> = {
  base: {
    "0": "m-0",
    "100": "m-100",
    "200": "m-200",
    "300": "m-300",
    "400": "m-400",
    "500": "m-500",
    "600": "m-600",
    "700": "m-700",
    "800": "m-800",
    "900": "m-900",
  },
  "sm": {
    "0": "sm:m-0",
    "100": "sm:m-100",
    "200": "sm:m-200",
    "300": "sm:m-300",
    "400": "sm:m-400",
    "500": "sm:m-500",
    "600": "sm:m-600",
    "700": "sm:m-700",
    "800": "sm:m-800",
    "900": "sm:m-900",
  },
  "md": {
    "0": "md:m-0",
    "100": "md:m-100",
    "200": "md:m-200",
    "300": "md:m-300",
    "400": "md:m-400",
    "500": "md:m-500",
    "600": "md:m-600",
    "700": "md:m-700",
    "800": "md:m-800",
    "900": "md:m-900",
  },
  "lg": {
    "0": "lg:m-0",
    "100": "lg:m-100",
    "200": "lg:m-200",
    "300": "lg:m-300",
    "400": "lg:m-400",
    "500": "lg:m-500",
    "600": "lg:m-600",
    "700": "lg:m-700",
    "800": "lg:m-800",
    "900": "lg:m-900",
  },
  "xl": {
    "0": "xl:m-0",
    "100": "xl:m-100",
    "200": "xl:m-200",
    "300": "xl:m-300",
    "400": "xl:m-400",
    "500": "xl:m-500",
    "600": "xl:m-600",
    "700": "xl:m-700",
    "800": "xl:m-800",
    "900": "xl:m-900",
  },
  "2xl": {
    "0": "2xl:m-0",
    "100": "2xl:m-100",
    "200": "2xl:m-200",
    "300": "2xl:m-300",
    "400": "2xl:m-400",
    "500": "2xl:m-500",
    "600": "2xl:m-600",
    "700": "2xl:m-700",
    "800": "2xl:m-800",
    "900": "2xl:m-900",
  },
};
// ...
export const mtMap: BaseClassMap<Spacing> = {
  "0": "mt-0",
  "100": "mt-100",
  "200": "mt-200",
  "300": "mt-300",
  "400": "mt-400",
  "500": "mt-500",
  "600": "mt-600",
  "700": "mt-700",
  "800": "mt-800",
  "900": "mt-900",
};

export const mrMap: BaseClassMap<Spacing> = {
  "0": "mr-0",
  "100": "mr-100",
  "200": "mr-200",
  "300": "mr-300",
  "400": "mr-400",
  "500": "mr-500",
  "600": "mr-600",
  "700": "mr-700",
  "800": "mr-800",
  "900": "mr-900",
};

export const mbMap: BaseClassMap<Spacing> = {
  "0": "mb-0",
  "100": "mb-100",
  "200": "mb-200",
  "300": "mb-300",
  "400": "mb-400",
  "500": "mb-500",
  "600": "mb-600",
  "700": "mb-700",
  "800": "mb-800",
  "900": "mb-900",
};

export const mlMap: BaseClassMap<Spacing> = {
  "0": "ml-0",
  "100": "ml-100",
  "200": "ml-200",
  "300": "ml-300",
  "400": "ml-400",
  "500": "ml-500",
  "600": "ml-600",
  "700": "ml-700",
  "800": "ml-800",
  "900": "ml-900",
};

export const gapMap: ResponsiveClassMap<Spacing> = {
  base: {
    "0": "gap-0",
    "100": "gap-100",
    "200": "gap-200",
    "300": "gap-300",
    "400": "gap-400",
    "500": "gap-500",
    "600": "gap-600",
    "700": "gap-700",
    "800": "gap-800",
    "900": "gap-900",
  },
  "sm": {
    "0": "sm:gap-0",
    "100": "sm:gap-100",
    "200": "sm:gap-200",
    "300": "sm:gap-300",
    "400": "sm:gap-400",
    "500": "sm:gap-500",
    "600": "sm:gap-600",
    "700": "sm:gap-700",
    "800": "sm:gap-800",
    "900": "sm:gap-900",
  },
  "md": {
    "0": "md:gap-0",
    "100": "md:gap-100",
    "200": "md:gap-200",
    "300": "md:gap-300",
    "400": "md:gap-400",
    "500": "md:gap-500",
    "600": "md:gap-600",
    "700": "md:gap-700",
    "800": "md:gap-800",
    "900": "md:gap-900",
  },
  "lg": {
    "0": "lg:gap-0",
    "100": "lg:gap-100",
    "200": "lg:gap-200",
    "300": "lg:gap-300",
    "400": "lg:gap-400",
    "500": "lg:gap-500",
    "600": "lg:gap-600",
    "700": "lg:gap-700",
    "800": "lg:gap-800",
    "900": "lg:gap-900",
  },
  "xl": {
    "0": "xl:gap-0",
    "100": "xl:gap-100",
    "200": "xl:gap-200",
    "300": "xl:gap-300",
    "400": "xl:gap-400",
    "500": "xl:gap-500",
    "600": "xl:gap-600",
    "700": "xl:gap-700",
    "800": "xl:gap-800",
    "900": "xl:gap-900",
  },
  "2xl": {
    "0": "2xl:gap-0",
    "100": "2xl:gap-100",
    "200": "2xl:gap-200",
    "300": "2xl:gap-300",
    "400": "2xl:gap-400",
    "500": "2xl:gap-500",
    "600": "2xl:gap-600",
    "700": "2xl:gap-700",
    "800": "2xl:gap-800",
    "900": "2xl:gap-900",
  },
};

export const gapXMap: ResponsiveClassMap<Spacing> = {
  base: {
    "0": "gap-x-0",
    "100": "gap-x-100",
    "200": "gap-x-200",
    "300": "gap-x-300",
    "400": "gap-x-400",
    "500": "gap-x-500",
    "600": "gap-x-600",
    "700": "gap-x-700",
    "800": "gap-x-800",
    "900": "gap-x-900",
  },
  "sm": {
    "0": "sm:gap-x-0",
    "100": "sm:gap-x-100",
    "200": "sm:gap-x-200",
    "300": "sm:gap-x-300",
    "400": "sm:gap-x-400",
    "500": "sm:gap-x-500",
    "600": "sm:gap-x-600",
    "700": "sm:gap-x-700",
    "800": "sm:gap-x-800",
    "900": "sm:gap-x-900",
  },
  "md": {
    "0": "md:gap-x-0",
    "100": "md:gap-x-100",
    "200": "md:gap-x-200",
    "300": "md:gap-x-300",
    "400": "md:gap-x-400",
    "500": "md:gap-x-500",
    "600": "md:gap-x-600",
    "700": "md:gap-x-700",
    "800": "md:gap-x-800",
    "900": "md:gap-x-900",
  },
  "lg": {
    "0": "lg:gap-x-0",
    "100": "lg:gap-x-100",
    "200": "lg:gap-x-200",
    "300": "lg:gap-x-300",
    "400": "lg:gap-x-400",
    "500": "lg:gap-x-500",
    "600": "lg:gap-x-600",
    "700": "lg:gap-x-700",
    "800": "lg:gap-x-800",
    "900": "lg:gap-x-900",
  },
  "xl": {
    "0": "xl:gap-x-0",
    "100": "xl:gap-x-100",
    "200": "xl:gap-x-200",
    "300": "xl:gap-x-300",
    "400": "xl:gap-x-400",
    "500": "xl:gap-x-500",
    "600": "xl:gap-x-600",
    "700": "xl:gap-x-700",
    "800": "xl:gap-x-800",
    "900": "xl:gap-x-900",
  },
  "2xl": {
    "0": "2xl:gap-x-0",
    "100": "2xl:gap-x-100",
    "200": "2xl:gap-x-200",
    "300": "2xl:gap-x-300",
    "400": "2xl:gap-x-400",
    "500": "2xl:gap-x-500",
    "600": "2xl:gap-x-600",
    "700": "2xl:gap-x-700",
    "800": "2xl:gap-x-800",
    "900": "2xl:gap-x-900",
  },
};

export const gapYMap: ResponsiveClassMap<Spacing> = {
  base: {
    "0": "gap-y-0",
    "100": "gap-y-100",
    "200": "gap-y-200",
    "300": "gap-y-300",
    "400": "gap-y-400",
    "500": "gap-y-500",
    "600": "gap-y-600",
    "700": "gap-y-700",
    "800": "gap-y-800",
    "900": "gap-y-900",
  },
  "sm": {
    "0": "sm:gap-y-0",
    "100": "sm:gap-y-100",
    "200": "sm:gap-y-200",
    "300": "sm:gap-y-300",
    "400": "sm:gap-y-400",
    "500": "sm:gap-y-500",
    "600": "sm:gap-y-600",
    "700": "sm:gap-y-700",
    "800": "sm:gap-y-800",
    "900": "sm:gap-y-900",
  },
  "md": {
    "0": "md:gap-y-0",
    "100": "md:gap-y-100",
    "200": "md:gap-y-200",
    "300": "md:gap-y-300",
    "400": "md:gap-y-400",
    "500": "md:gap-y-500",
    "600": "md:gap-y-600",
    "700": "md:gap-y-700",
    "800": "md:gap-y-800",
    "900": "md:gap-y-900",
  },
  "lg": {
    "0": "lg:gap-y-0",
    "100": "lg:gap-y-100",
    "200": "lg:gap-y-200",
    "300": "lg:gap-y-300",
    "400": "lg:gap-y-400",
    "500": "lg:gap-y-500",
    "600": "lg:gap-y-600",
    "700": "lg:gap-y-700",
    "800": "lg:gap-y-800",
    "900": "lg:gap-y-900",
  },
  "xl": {
    "0": "xl:gap-y-0",
    "100": "xl:gap-y-100",
    "200": "xl:gap-y-200",
    "300": "xl:gap-y-300",
    "400": "xl:gap-y-400",
    "500": "xl:gap-y-500",
    "600": "xl:gap-y-600",
    "700": "xl:gap-y-700",
    "800": "xl:gap-y-800",
    "900": "xl:gap-y-900",
  },
  "2xl": {
    "0": "2xl:gap-y-0",
    "100": "2xl:gap-y-100",
    "200": "2xl:gap-y-200",
    "300": "2xl:gap-y-300",
    "400": "2xl:gap-y-400",
    "500": "2xl:gap-y-500",
    "600": "2xl:gap-y-600",
    "700": "2xl:gap-y-700",
    "800": "2xl:gap-y-800",
    "900": "2xl:gap-y-900",
  },
};
// ...
export const heightMap: ResponsiveClassMap<Height> = {
  base: {
    "auto": "h-auto",
    "full": "h-full",
    "screen": "h-screen",
    "svh": "h-svh",
    "dvh": "h-dvh",
  },
  "sm": {
    "auto": "sm:h-auto",
    "full": "sm:h-full",
    "screen": "sm:h-screen",
    "svh": "sm:h-svh",
    "dvh": "sm:h-dvh",
  },
  "md": {
    "auto": "md:h-auto",
    "full": "md:h-full",
    "screen": "md:h-screen",
    "svh": "md:h-svh",
    "dvh": "md:h-dvh",
  },
  "lg": {
    "auto": "lg:h-auto",
    "full": "lg:h-full",
    "screen": "lg:h-screen",
    "svh": "lg:h-svh",
    "dvh": "lg:h-dvh",
  },
  "xl": {
    "auto": "xl:h-auto",
    "full": "xl:h-full",
    "screen": "xl:h-screen",
    "svh": "xl:h-svh",
    "dvh": "xl:h-dvh",
  },
  "2xl": {
    "auto": "2xl:h-auto",
    "full": "2xl:h-full",
    "screen": "2xl:h-screen",
    "svh": "2xl:h-svh",
    "dvh": "2xl:h-dvh",
  },
};

export const minHeightMap: ResponsiveClassMap<MinHeight> = {
  base: {
    "none": "min-h-auto",
    "full": "min-h-full",
    "screen": "min-h-screen",
    "svh": "min-h-svh",
    "dvh": "min-h-dvh",
  },
  "sm": {
    "none": "sm:min-h-auto",
    "full": "sm:min-h-full",
    "screen": "sm:min-h-screen",
    "svh": "sm:min-h-svh",
    "dvh": "sm:min-h-dvh",
  },
  "md": {
    "none": "md:min-h-auto",
    "full": "md:min-h-full",
    "screen": "md:min-h-screen",
    "svh": "md:min-h-svh",
    "dvh": "md:min-h-dvh",
  },
  "lg": {
    "none": "lg:min-h-auto",
    "full": "lg:min-h-full",
    "screen": "lg:min-h-screen",
    "svh": "lg:min-h-svh",
    "dvh": "lg:min-h-dvh",
  },
  "xl": {
    "none": "xl:min-h-auto",
    "full": "xl:min-h-full",
    "screen": "xl:min-h-screen",
    "svh": "xl:min-h-svh",
    "dvh": "xl:min-h-dvh",
  },
  "2xl": {
    "none": "2xl:min-h-auto",
    "full": "2xl:min-h-full",
    "screen": "2xl:min-h-screen",
    "svh": "2xl:min-h-svh",
    "dvh": "2xl:min-h-dvh",
  },
};

export const directionMap: ResponsiveClassMap<Direction> = {
  base: {
    "row": "flex-row",
    "row-reverse": "flex-row-reverse",
    "col": "flex-col",
    "col-reverse": "flex-col-reverse",
  },
  "sm": {
    "row": "sm:flex-row",
    "row-reverse": "sm:flex-row-reverse",
    "col": "sm:flex-col",
    "col-reverse": "sm:flex-col-reverse",
  },
  "md": {
    "row": "md:flex-row",
    "row-reverse": "md:flex-row-reverse",
    "col": "md:flex-col",
    "col-reverse": "md:flex-col-reverse",
  },
  "lg": {
    "row": "lg:flex-row",
    "row-reverse": "lg:flex-row-reverse",
    "col": "lg:flex-col",
    "col-reverse": "lg:flex-col-reverse",
  },
  "xl": {
    "row": "xl:flex-row",
    "row-reverse": "xl:flex-row-reverse",
    "col": "xl:flex-col",
    "col-reverse": "xl:flex-col-reverse",
  },
  "2xl": {
    "row": "2xl:flex-row",
    "row-reverse": "2xl:flex-row-reverse",
    "col": "2xl:flex-col",
    "col-reverse": "2xl:flex-col-reverse",
  },
};
// ...
export const justifyMap: ResponsiveClassMap<Justify> = {
  base: {
    "start": "justify-start",
    "center": "justify-center",
    "end": "justify-end",
    "between": "justify-between",
    "around": "justify-around",
    "evenly": "justify-evenly",
  },
  "sm": {
    "start": "sm:justify-start",
    "center": "sm:justify-center",
    "end": "sm:justify-end",
    "between": "sm:justify-between",
    "around": "sm:justify-around",
    "evenly": "sm:justify-evenly",
  },
  "md": {
    "start": "md:justify-start",
    "center": "md:justify-center",
    "end": "md:justify-end",
    "between": "md:justify-between",
    "around": "md:justify-around",
    "evenly": "md:justify-evenly",
  },
  "lg": {
    "start": "lg:justify-start",
    "center": "lg:justify-center",
    "end": "lg:justify-end",
    "between": "lg:justify-between",
    "around": "lg:justify-around",
    "evenly": "lg:justify-evenly",
  },
  "xl": {
    "start": "xl:justify-start",
    "center": "xl:justify-center",
    "end": "xl:justify-end",
    "between": "xl:justify-between",
    "around": "xl:justify-around",
    "evenly": "xl:justify-evenly",
  },
  "2xl": {
    "start": "2xl:justify-start",
    "center": "2xl:justify-center",
    "end": "2xl:justify-end",
    "between": "2xl:justify-between",
    "around": "2xl:justify-around",
    "evenly": "2xl:justify-evenly",
  },
};